+++ src/onegov.election_day/onegov/election_day/formats/election/wabsti/__init__.py
@@ -9,7 +9,8 @@ from onegov.election_day.formats.election.wabsti.proporz import (
 def import_file(municipalities, election, file, mimetype,
                 connections_file=None, connections_mimetype=None,
                 elected_file=None, elected_mimetype=None,
-                statistics_file=None, statistics_mimetype=None):
+                statistics_file=None, statistics_mimetype=None,
+                parties_file=None, parties_mimetype=None):
     """ Tries to import the given file (wabsti format).
 
     :return: A dictionary containing the status and a list of errors if any.
@@ -30,5 +31,6 @@ def import_file(municipalities, election, file, mimetype,
             municipalities, election, file, mimetype,
             connections_file, connections_mimetype,
             elected_file, elected_mimetype,
-            statistics_file, statistics_mimetype
+            statistics_file, statistics_mimetype,
+            parties_file, parties_mimetype
         )